That didn't quite work because we weren't generating schemes.

+++ code/source/010_3D_Cameras/010_3D_Cameras.c
@@ -832,7 +832,7 @@ SDL_GPUTextureFormat GetSupportedDepthFormat()
 }
 
 
-SDL_GPUBuffer* CreateAndUploadBuffer(const void* aData, size_t aSize, SDL_GPUBufferUsageFlags aUsage)
+SDL_GPUBuffer* CreateAndUploadBuffer(const void* aData, Uint32 aSize, SDL_GPUBufferUsageFlags aUsage)
 {
   SDL_GPUBufferCreateInfo bufferCreateInfo;
   SDL_zero(bufferCreateInfo);
